Add files via upload
GestureDetectoCode.py is the program I coded to run the task. It contains a variable "usingRaspberryPi" that should be set to True when running on a real Raspberry Pi and False to use it with the simulator code.
UltrasonicRangerInterface.py contains code that accesses the rangefinder on the Raspberry Pi.
SonarSim.py contains code that pretends to be accessing a rangefinder and sends values that simulate the hand gestures you need to write code to recognise.
You can run GestureDetector.py on any computer using the simulator code, or on a Raspberry Pi equipped with the rangefinder.